# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: ERP Barrio API
  version: 0.3.8
  description: Contrato inicial del API del MVP ERP.
servers:
  - url: http://localhost:8000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
...
    ProductCreateRequest:
      type: object
      required: [sku, name, price]
      properties:
        sku:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
        price:
          type: number
          exclusiveMinimum: 0
    ProductUpdateRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        price:
          type: number
          exclusiveMinimum: 0
    ProductResponse:
      type: object
      required: [id, sku, name, price]
      properties:
        id:
          type: string
        sku:
          type: string
        name:
          type: string
        price:
          type: number
    ProductListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProductResponse'
...
paths:
...
  /products:
    get:
      tags: [products]
      summary: Lista productos del catálogo
      operationId: getProducts
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Listado de productos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
        '401':
          description: Token ausente/inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Rol insuficiente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags: [products]
      summary: Crea un producto en catálogo
      operationId: postProducts
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreateRequest'
      responses:
        '201':
          description: Producto creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '401':
          description: Token ausente/inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Rol insuficiente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: SKU duplicado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /products/{product_id}:
    parameters:
      - in: path
        name: product_id
        required: true
        schema:
          type: string
    get:
      tags: [products]
      summary: Obtiene un producto por identificador
      operationId: getProductById
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Producto encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '401':
          description: Token ausente/inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Rol insuficiente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Producto no existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags: [products]
      summary: Actualiza campos de un producto
      operationId: patchProductById
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdateRequest'
      responses:
        '200':
          description: Producto actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '401':
          description: Token ausente/inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Rol insuficiente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Producto no existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [products]
      summary: Elimina un producto del catálogo
      operationId: deleteProductById
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Producto eliminado
        '401':
          description: Token ausente/inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Rol insuficiente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Producto no existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...